+++ CHANGELOG.md
@@ -15,6 +15,8 @@
     - `ElasticsearchEx.Api.Search.Core.get_scroll/2`
     - `ElasticsearchEx.Api.Search.Core.clear_scroll/2`
   * Added a new module `ElasticsearchEx.Ndjson` to manipulate NDJSON
+  * Added new functions related to `multi_search`:
+    - `ElasticsearchEx.Api.Search.Core.multi_search/2`
 
 * **Changes:**
   * Added Github actions matrix to test different versions